+++ packages/riverpod/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased build
 
-Bump minimum `meta` version
+- Bump minimum `meta` version
+- Added `AsyncValue.retrying`, to check when a retry is scheduled or pending
 
 ## 3.0.0-dev.16 - 2025-06-20