+++ packages/riverpod/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bump minimum `meta` version
 - Added `AsyncValue.retrying`, to check when a retry is scheduled or pending
 - Exposed the default retry implementation (`ProviderContainer.defaultRetry`)
+- Offline's Storage now is `base` and requires overriding `deleteOutOfDate`
 
 ## 3.0.0-dev.16 - 2025-06-20