+++ packages/riverpod/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased build
+## 3.0.0-dev.17 - 2025-08-01
 
 - Added `MutationState.isPending/isIdle/hasError/isSuccess`
 - fixes various "pause" issues